created FluidContainer folder and component
