--- front-end/src/components/Common/index.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import {Container, Row, Col} from 'react-bootstrap'
-
-function index() {
-    return (
-        
-            <Container fluid>
-                <Row>
-                    <Col>1 of 1</Col>
-                </Row>
-            </Container>
-        
-    )
-}
-
-export default index
